backend/requirements.txt
yfinance
langchain-pinecone
openai
python-dotenv
langchain-community
sentence-transformers
fastapi
uvicorn
pinecone-client
langchain
pydantic
